feat: create Raffle SmartContract

+++ hardhat/contracts/Lock.sol
@@ -5,30 +5,30 @@ pragma solidity ^0.8.9;
 // import "hardhat/console.sol";
 
 contract Lock {
-    uint public unlockTime;
-    address payable public owner;
+  uint256 public unlockTime;
+  address payable public owner;
 
-    event Withdrawal(uint amount, uint when);
+  event Withdrawal(uint256 amount, uint256 when);
 
-    constructor(uint _unlockTime) payable {
-        require(
-            block.timestamp < _unlockTime,
-            "Unlock time should be in the future"
-        );
+  constructor(uint256 _unlockTime) payable {
+    require(
+      block.timestamp < _unlockTime,
+      'Unlock time should be in the future'
+    );
 
-        unlockTime = _unlockTime;
-        owner = payable(msg.sender);
-    }
+    unlockTime = _unlockTime;
+    owner = payable(msg.sender);
+  }
 
-    function withdraw() public {
-        // Uncomment this line to print a log in your terminal
-        // console.log("Unlock time is %o and block timestamp is %o", unlockTime, block.timestamp);
+  function withdraw() public {
+    // Uncomment this line to print a log in your terminal
+    // console.log("Unlock time is %o and block timestamp is %o", unlockTime, block.timestamp);
 
-        require(block.timestamp >= unlockTime, "You can't withdraw yet");
-        require(msg.sender == owner, "You aren't the owner");
+    require(block.timestamp >= unlockTime, "You can't withdraw yet");
+    require(msg.sender == owner, "You aren't the owner");
 
-        emit Withdrawal(address(this).balance, block.timestamp);
+    emit Withdrawal(address(this).balance, block.timestamp);
 
-        owner.transfer(address(this).balance);
-    }
+    owner.transfer(address(this).balance);
+  }
 }
